requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
chromadb>=0.4.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0